+++ plugin/src/commonInterfaces.ts
@@ -1,0 +1,17 @@
+import * as BabelCore from '@babel/core';
+
+interface ReanimatedPluginPass {
+  file: BabelCore.BabelFile;
+  key: string;
+  opts: {
+    relativeSourceLocation?: boolean;
+    disableInlineStylesWarning?: boolean;
+  };
+  cwd: string;
+  filename: string | undefined;
+  get(key: unknown): unknown;
+  set(key: unknown, value: unknown): void;
+  [key: string]: unknown;
+}
+
+export { ReanimatedPluginPass };
