+++ plugin/src/makeWorklet.ts
@@ -1,0 +1,461 @@
+import * as BabelCore from '@babel/core';
+import * as BabelTypes from '@babel/types';
+import generate from '@babel/generator';
+import traverse from '@babel/traverse';
+import { transformSync } from '@babel/core';
+import * as fs from 'fs';
+import * as convertSourceMap from 'convert-source-map';
+import { ReanimatedPluginPass } from './commonInterfaces';
+import { shouldGenerateSourceMap, hash, isRelease } from './commonFunctions';
+import { globals } from './commonObjects';
+
+function buildWorkletString(
+  t: typeof BabelCore.types,
+  fun: BabelCore.types.File,
+  closureVariables: Array<BabelTypes.Identifier>,
+  name: string,
+  inputMap: BabelCore.BabelFileResult['map']
+): Array<string | null | undefined> {
+  function prependClosureVariablesIfNecessary() {
+    const closureDeclaration = t.variableDeclaration('const', [
+      t.variableDeclarator(
+        t.objectPattern(
+          closureVariables.map((variable) =>
+            t.objectProperty(
+              t.identifier(variable.name),
+              t.identifier(variable.name),
+              false,
+              true
+            )
+          )
+        ),
+        t.memberExpression(t.thisExpression(), t.identifier('_closure'))
+      ),
+    ]);
+
+    function prependClosure(
+      path: BabelCore.NodePath<
+        | BabelTypes.FunctionDeclaration
+        | BabelTypes.FunctionExpression
+        | BabelTypes.ArrowFunctionExpression
+        | BabelTypes.ObjectMethod
+      >
+    ) {
+      if (closureVariables.length === 0 || path.parent.type !== 'Program') {
+        return;
+      }
+
+      if (!BabelTypes.isExpression(path.node.body))
+        path.node.body.body.unshift(closureDeclaration);
+    }
+
+    function prependRecursiveDeclaration(
+      path: BabelCore.NodePath<
+        | BabelTypes.FunctionDeclaration
+        | BabelTypes.FunctionExpression
+        | BabelTypes.ArrowFunctionExpression
+        | BabelTypes.ObjectMethod
+      >
+    ) {
+      if (
+        path.parent.type === 'Program' &&
+        !BabelTypes.isArrowFunctionExpression(path.node) &&
+        !BabelTypes.isObjectMethod(path.node) &&
+        path.node.id &&
+        path.scope.parent
+      ) {
+        const hasRecursiveCalls =
+          path.scope.parent.bindings[path.node.id.name]?.references > 0;
+        if (hasRecursiveCalls) {
+          path.node.body.body.unshift(
+            t.variableDeclaration('const', [
+              t.variableDeclarator(
+                t.identifier(path.node.id.name),
+                t.memberExpression(t.thisExpression(), t.identifier('_recur'))
+              ),
+            ])
+          );
+        }
+      }
+    }
+
+    return {
+      visitor: {
+        'FunctionDeclaration|FunctionExpression|ArrowFunctionExpression|ObjectMethod':
+          (
+            path: BabelCore.NodePath<
+              | BabelTypes.FunctionDeclaration
+              | BabelTypes.FunctionExpression
+              | BabelTypes.ArrowFunctionExpression
+              | BabelTypes.ObjectMethod
+            >
+          ) => {
+            prependClosure(path);
+            prependRecursiveDeclaration(path);
+          },
+      },
+    };
+  }
+
+  const draftExpression = (fun.program.body.find((obj) =>
+    BabelTypes.isFunctionDeclaration(obj)
+  ) ||
+    fun.program.body.find((obj) => BabelTypes.isExpressionStatement(obj)) ||
+    undefined) as
+    | BabelTypes.FunctionDeclaration
+    | BabelTypes.ExpressionStatement
+    | undefined;
+
+  if (!draftExpression) throw new Error("'draftExpression' is not defined\n");
+
+  const expression = BabelTypes.isFunctionDeclaration(draftExpression)
+    ? draftExpression
+    : draftExpression.expression;
+
+  if (!('params' in expression && BabelTypes.isBlockStatement(expression.body)))
+    throw new Error(
+      "'expression' doesn't have property 'params' or 'expression.body' is not a BlockStatmenent\n'"
+    );
+
+  const workletFunction = BabelTypes.functionExpression(
+    BabelTypes.identifier(name),
+    expression.params,
+    expression.body
+  );
+
+  const code = generate(workletFunction).code;
+
+  if (!inputMap) throw new Error("'inputMap' is not defined");
+
+  const includeSourceMap = shouldGenerateSourceMap();
+
+  if (includeSourceMap) {
+    // Clear contents array (should be empty anyways)
+    inputMap.sourcesContent = [];
+    // Include source contents in source map, because Flipper/iframe is not
+    // allowed to read files from disk.
+    for (const sourceFile of inputMap.sources) {
+      inputMap.sourcesContent.push(
+        fs.readFileSync(sourceFile).toString('utf-8')
+      );
+    }
+  }
+
+  const transformed = transformSync(code, {
+    plugins: [prependClosureVariablesIfNecessary()],
+    compact: !includeSourceMap,
+    sourceMaps: includeSourceMap,
+    inputSourceMap: inputMap,
+    ast: false,
+    babelrc: false,
+    configFile: false,
+    comments: false,
+  });
+
+  if (!transformed) throw new Error('transformed is null!\n');
+
+  let sourceMap;
+  if (includeSourceMap) {
+    sourceMap = convertSourceMap.fromObject(transformed.map).toObject();
+    // sourcesContent field contains a full source code of the file which contains the worklet
+    // and is not needed by the source map interpreter in order to symbolicate a stack trace.
+    // Therefore, we remove it to reduce the bandwith and avoid sending it potentially multiple times
+    // in files that contain multiple worklets. Along with sourcesContent.
+    delete sourceMap.sourcesContent;
+  }
+
+  return [transformed.code, JSON.stringify(sourceMap)];
+}
+
+function makeWorkletName(
+  t: typeof BabelCore.types,
+  fun: BabelCore.NodePath<
+    | BabelTypes.FunctionDeclaration
+    | BabelTypes.FunctionExpression
+    | BabelTypes.ObjectMethod
+    | BabelTypes.ArrowFunctionExpression
+  >
+): string {
+  if (t.isObjectMethod(fun.node) && 'name' in fun.node.key) {
+    return fun.node.key.name;
+  }
+  if (t.isFunctionDeclaration(fun.node) && fun.node.id) {
+    return fun.node.id.name;
+  }
+  if (
+    BabelTypes.isFunctionExpression(fun.node) &&
+    BabelTypes.isIdentifier(fun.node.id)
+  ) {
+    return fun.node.id.name;
+  }
+  return 'anonymous'; // fallback for ArrowFunctionExpression and unnamed FunctionExpression
+}
+
+function makeWorklet(
+  t: typeof BabelCore.types,
+  fun: BabelCore.NodePath<
+    | BabelTypes.FunctionDeclaration
+    | BabelTypes.FunctionExpression
+    | BabelTypes.ObjectMethod
+    | BabelTypes.ArrowFunctionExpression
+  >,
+  state: ReanimatedPluginPass
+): BabelTypes.FunctionExpression {
+  // Returns a new FunctionExpression which is a workletized version of provided
+  // FunctionDeclaration, FunctionExpression, ArrowFunctionExpression or ObjectMethod.
+
+  const functionName = makeWorkletName(t, fun);
+
+  const closure = new Map<string, BabelTypes.Identifier>();
+
+  // remove 'worklet'; directive before generating string
+  fun.traverse({
+    DirectiveLiteral(path) {
+      if (path.node.value === 'worklet' && path.getFunctionParent() === fun) {
+        path.parentPath.remove();
+      }
+    },
+  });
+
+  // We use copy because some of the plugins don't update bindings and
+  // some even break them
+  if (!state.file.opts.filename)
+    throw new Error("'state.file.opts.filename' is undefined\n");
+
+  const codeObject = generate(fun.node, {
+    sourceMaps: true,
+    sourceFileName: state.file.opts.filename,
+  });
+
+  // We need to add a newline at the end, because there could potentially be a
+  // comment after the function that gets included here, and then the closing
+  // bracket would become part of the comment thus resulting in an error, since
+  // there is a missing closing bracket.
+  const code =
+    '(' + (t.isObjectMethod(fun) ? 'function ' : '') + codeObject.code + '\n)';
+
+  const transformed = transformSync(code, {
+    filename: state.file.opts.filename,
+    presets: ['@babel/preset-typescript'],
+    plugins: [
+      '@babel/plugin-transform-shorthand-properties',
+      '@babel/plugin-transform-arrow-functions',
+      '@babel/plugin-proposal-optional-chaining',
+      '@babel/plugin-proposal-nullish-coalescing-operator',
+      ['@babel/plugin-transform-template-literals', { loose: true }],
+    ],
+    ast: true,
+    babelrc: false,
+    configFile: false,
+    inputSourceMap: codeObject.map,
+  });
+
+  if (!transformed || !transformed.ast)
+    throw new Error("'transformed' or 'transformed.ast' is undefined\n");
+
+  traverse(transformed.ast, {
+    Identifier(path) {
+      if (!path.isReferencedIdentifier()) return;
+      const name = path.node.name;
+      if (
+        globals.has(name) ||
+        (!BabelTypes.isArrowFunctionExpression(fun.node) &&
+          !BabelTypes.isObjectMethod(fun.node) &&
+          fun.node.id &&
+          fun.node.id.name === name)
+      ) {
+        return;
+      }
+
+      const parentNode = path.parent;
+
+      if (
+        parentNode.type === 'MemberExpression' &&
+        parentNode.property === path.node &&
+        !parentNode.computed
+      ) {
+        return;
+      }
+
+      if (
+        parentNode.type === 'ObjectProperty' &&
+        path.parentPath.parent.type === 'ObjectExpression' &&
+        path.node !== parentNode.value
+      ) {
+        return;
+      }
+
+      let currentScope = path.scope;
+
+      while (currentScope != null) {
+        if (currentScope.bindings[name] != null) {
+          return;
+        }
+        currentScope = currentScope.parent;
+      }
+      closure.set(name, path.node);
+    },
+  });
+
+  const variables = Array.from(closure.values());
+
+  const privateFunctionId = t.identifier('_f');
+  const clone = t.cloneNode(fun.node);
+  const funExpression = BabelTypes.isBlockStatement(clone.body)
+    ? BabelTypes.functionExpression(null, clone.params, clone.body)
+    : clone;
+
+  const [funString, sourceMapString] = buildWorkletString(
+    t,
+    transformed.ast,
+    variables,
+    functionName,
+    transformed.map
+  );
+  if (!funString) throw new Error("'funString' is not defined\n");
+  const workletHash = hash(funString);
+
+  let location = state.file.opts.filename;
+  if (state.opts.relativeSourceLocation) {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const path = require('path');
+    location = path.relative(state.cwd, location);
+  }
+
+  let lineOffset = 1;
+  if (closure.size > 0) {
+    // When worklet captures some variables, we append closure destructing at
+    // the beginning of the function body. This effectively results in line
+    // numbers shifting by the number of captured variables (size of the
+    // closure) + 2 (for the opening and closing brackets of the destruct
+    // statement)
+    lineOffset -= closure.size + 2;
+  }
+
+  const pathForStringDefinitions = fun.parentPath.isProgram()
+    ? fun
+    : (fun.findParent(
+        (path) =>
+          (path.parentPath as BabelCore.NodePath<BabelCore.Node>).isProgram() // this causes typescript error on Windows CI build
+      ) as BabelCore.NodePath<BabelCore.Node>); // this causes typescript error on Windows CI build
+
+  const initDataId = (
+    pathForStringDefinitions.parentPath as BabelCore.NodePath<BabelCore.Node>
+  ).scope // this causes typescript error on Windows CI build
+    .generateUidIdentifier(`worklet_${workletHash}_init_data`);
+
+  const initDataObjectExpression = t.objectExpression([
+    t.objectProperty(t.identifier('code'), t.stringLiteral(funString)),
+    t.objectProperty(t.identifier('location'), t.stringLiteral(location)),
+  ]);
+
+  if (sourceMapString) {
+    initDataObjectExpression.properties.push(
+      t.objectProperty(
+        t.identifier('sourceMap'),
+        t.stringLiteral(sourceMapString)
+      )
+    );
+  }
+
+  pathForStringDefinitions.insertBefore(
+    t.variableDeclaration('const', [
+      t.variableDeclarator(initDataId, initDataObjectExpression),
+    ])
+  );
+
+  if (
+    BabelTypes.isFunctionDeclaration(funExpression) ||
+    BabelTypes.isObjectMethod(funExpression)
+  )
+    throw new Error(
+      "'funExpression' is either FunctionDeclaration or ObjectMethod and cannot be used in variableDeclaration\n"
+    );
+
+  const statements: Array<
+    | BabelTypes.VariableDeclaration
+    | BabelTypes.ExpressionStatement
+    | BabelTypes.ReturnStatement
+  > = [
+    t.variableDeclaration('const', [
+      t.variableDeclarator(privateFunctionId, funExpression),
+    ]),
+    t.expressionStatement(
+      t.assignmentExpression(
+        '=',
+        t.memberExpression(privateFunctionId, t.identifier('_closure'), false),
+        t.objectExpression(
+          variables.map((variable) =>
+            t.objectProperty(t.identifier(variable.name), variable, false, true)
+          )
+        )
+      )
+    ),
+    t.expressionStatement(
+      t.assignmentExpression(
+        '=',
+        t.memberExpression(
+          privateFunctionId,
+          t.identifier('__initData'),
+          false
+        ),
+        initDataId
+      )
+    ),
+    t.expressionStatement(
+      t.assignmentExpression(
+        '=',
+        t.memberExpression(
+          privateFunctionId,
+          t.identifier('__workletHash'),
+          false
+        ),
+        t.numericLiteral(workletHash)
+      )
+    ),
+  ];
+
+  if (!isRelease()) {
+    statements.unshift(
+      t.variableDeclaration('const', [
+        t.variableDeclarator(
+          t.identifier('_e'),
+          t.arrayExpression([
+            t.newExpression(
+              t.memberExpression(t.identifier('global'), t.identifier('Error')),
+              []
+            ),
+            t.numericLiteral(lineOffset),
+            t.numericLiteral(-27), // the placement of opening bracket after Exception in line that defined '_e' variable
+          ])
+        ),
+      ])
+    );
+    statements.push(
+      t.expressionStatement(
+        t.assignmentExpression(
+          '=',
+          t.memberExpression(
+            privateFunctionId,
+            t.identifier('__stackDetails'),
+            false
+          ),
+          t.identifier('_e')
+        )
+      )
+    );
+  }
+
+  statements.push(t.returnStatement(privateFunctionId));
+
+  const newFun = t.functionExpression(
+    // !BabelTypes.isArrowFunctionExpression(fun.node) ? fun.node.id : undefined, // [TO DO] --- this never worked
+    undefined,
+    [],
+    t.blockStatement(statements)
+  );
+
+  return newFun;
+}
+
+export { makeWorklet };
