+++ plugin/src/processIfWorkletNode.ts
@@ -1,0 +1,43 @@
+import * as BabelCore from '@babel/core';
+import * as BabelTypes from '@babel/types';
+import { ReanimatedPluginPass } from './commonInterfaces';
+import { processWorkletFunction } from './processWorkletFunction';
+
+function processIfWorkletNode(
+  t: typeof BabelCore.types,
+  fun: BabelCore.NodePath<
+    | BabelTypes.FunctionDeclaration
+    | BabelTypes.FunctionExpression
+    | BabelTypes.ArrowFunctionExpression
+  >,
+  state: ReanimatedPluginPass
+) {
+  fun.traverse({
+    DirectiveLiteral(path) {
+      const value = path.node.value;
+      if (
+        value === 'worklet' &&
+        path.getFunctionParent() === fun &&
+        BabelTypes.isBlockStatement(fun.node.body)
+      ) {
+        // make sure "worklet" is listed among directives for the fun
+        // this is necessary as because of some bug, babel will attempt to
+        // process replaced function if it is nested inside another function
+        const directives = fun.node.body.directives;
+        if (
+          directives &&
+          directives.length > 0 &&
+          directives.some(
+            (directive) =>
+              t.isDirectiveLiteral(directive.value) &&
+              directive.value.value === 'worklet'
+          )
+        ) {
+          processWorkletFunction(t, fun, state);
+        }
+      }
+    },
+  });
+}
+
+export { processIfWorkletNode };
