+++ plugin/src/processWorkletFunction.ts
@@ -1,0 +1,44 @@
+import * as BabelCore from '@babel/core';
+import * as BabelTypes from '@babel/types';
+import { ReanimatedPluginPass } from './commonInterfaces';
+import { makeWorklet } from './makeWorklet';
+
+function processWorkletFunction(
+  t: typeof BabelCore.types,
+  fun: BabelCore.NodePath<
+    | BabelTypes.FunctionDeclaration
+    | BabelTypes.FunctionExpression
+    | BabelTypes.ArrowFunctionExpression
+  >,
+  state: ReanimatedPluginPass
+) {
+  // Replaces FunctionDeclaration, FunctionExpression or ArrowFunctionExpression
+  // with a workletized version of itself.
+
+  if (!t.isFunctionParent(fun)) {
+    return;
+  }
+
+  const newFun = makeWorklet(t, fun, state);
+
+  const replacement = t.callExpression(newFun, []);
+
+  // we check if function needs to be assigned to variable declaration.
+  // This is needed if function definition directly in a scope. Some other ways
+  // where function definition can be used is for example with variable declaration:
+  // const ggg = function foo() { }
+  // ^ in such a case we don't need to define variable for the function
+  const needDeclaration =
+    t.isScopable(fun.parent) || t.isExportNamedDeclaration(fun.parent);
+  fun.replaceWith(
+    !BabelTypes.isArrowFunctionExpression(fun.node) &&
+      fun.node.id &&
+      needDeclaration
+      ? t.variableDeclaration('const', [
+          t.variableDeclarator(fun.node.id, replacement),
+        ])
+      : replacement
+  );
+}
+
+export { processWorkletFunction };
